Introduce ImageThumbnailer expect/actual for photo upload (Phase 4b)
Per explicit confirmation, introduces the expect/actual now rather than
deferring to the iOS phase.

expect class PlatformImageSource / ImageThumbnailer / expect fun
createImageThumbnailer() land in :core's commonMain. The Android actual
is a straight relocation of the existing ContentResolver/ImageDecoder/
MediaStore/Bitmap thumbnailing logic (PlatformImageSource is an actual
typealias for Uri, so existing Android call sites passing a Uri need no
change at all). createImageThumbnailer() follows the same GlobalContext
pattern as Phase 3B's getDatabaseBuilder(). The iOS actual (NSURL,
PHAsset, or raw ByteArray, depending on how the iOS photo-picker UI
ends up being built) is deferred to the iOS phase, since nothing
consumes it yet.

UploadPhoto moves to commonMain, now taking PlatformImageSource +
ImageThumbnailer instead of Uri + ContentResolver directly. Its DI
wiring (core/domain/user_data/di/Module.kt, which stays androidMain -
most of its other use cases are still Android-coupled) is updated to
call createImageThumbnailer() instead of reaching into
get<Context>().contentResolver directly.

Unverified in this sandbox - no Android SDK, Gradle 9.3.1 blocked by
network policy.
