--- core/src/androidMain/kotlin/com/hussienFahmy/core/domain/user_data/use_cases/UploadPhoto.kt
@@ -1,41 +1,0 @@
-package com.hussienfahmy.core.domain.user_data.use_cases
-
-import android.content.ContentResolver
-import android.graphics.Bitmap
-import android.graphics.ImageDecoder
-import android.net.Uri
-import android.os.Build
-import android.provider.MediaStore
-import com.hussienfahmy.core.domain.storage.repository.StorageRepository
-import com.hussienfahmy.core.domain.user_data.repository.UserDataRepository
-import kotlinx.coroutines.flow.first
-import java.io.ByteArrayOutputStream
-
-class UploadPhoto(
-    private val repository: UserDataRepository,
-    private val updatePhotoUrl: UpdatePhotoUrl,
-    private val storageRepository: StorageRepository,
-    private val contentResolver: ContentResolver,
-) {
-    suspend operator fun invoke(uri: Uri) {
-        val user = repository.userData.first() ?: return
-
-        val thumbnail = createThumbnail(uri)
-        val downloadUrl = storageRepository.uploadUserPhoto(user.id, thumbnail)
-
-        updatePhotoUrl(downloadUrl)
-    }
-
-    private fun createThumbnail(uri: Uri): ByteArray {
-        val bitmap = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-            val source = ImageDecoder.createSource(contentResolver, uri)
-            ImageDecoder.decodeBitmap(source)
-        } else {
-            MediaStore.Images.Media.getBitmap(contentResolver, uri)
-        }
-
-        val outputStream = ByteArrayOutputStream()
-        bitmap.compress(Bitmap.CompressFormat.JPEG, 60, outputStream)
-        return outputStream.toByteArray()
-    }
-}
